+++ client/src/components/Contact/contact.scss
@@ -1,24 +1,5 @@
 .contact {
-    display: flex;
+    text-align: center;
+    padding: 3rem;
 
-    .form {
-        flex: 1;
-    }
-
-    .info {
-        flex: 1;
-
-        .personal {
-            .personal-item {
-                display: flex;
-
-                .icon {
-                    padding-right: 20px;
-                }
-            }
-        }
-
-        .links {}
-
-    }
 }
